Initial Import
The first versioned files of the project. The MalDB module contains
the basic database io functions, the CleanText module contains
functions to clean up text for storage, and the WebGrab module contains
functions to get data from the internet. The webtest file performs
the basic test of grabbing a list and userid from the internet and
storing the appropriate entries in the database.