Allowed min rater restrictions to apply to arbitrary sets
The function to restrict anime to a set that has only contains
anime with a minimum number of raters now takes the set to restrict
from as an argument, rather than restricting from the full set
in the database. This allows different types of restrictions to
apply. Also, the restriction notice now only prints if the verbose
flag is set.