Added functions to get the list of u/v ids in a factor model
The getUIds and getVIds functions of FactorModel return the list
of uids and vids respectively. This is needed when the factor
model is saved, but the database gains new data between the time
the model is created and the time it is used. In that case, the
database may have more ids than is present in the model, so it
is necessary to get the ids present in the model instead.