Separated the model generation and recommendation
The RecommendAnime script used to generate a model and then used
it to recommend anime. It has been split to two different scripts,
CreateLinearFactorModel creates the model and RecommendAnime
loads the model and uses to recommend. This makes it easier to
make several recommendations based on the same model, and swap
models easily.