Changed stop criterion from error decrease to iteration limit
The stopping criterion of the regression loop of making a
linear factor model is now a max iteration limit. This was done
as error calculation wasn't quite correct and correct calculation
would take more time and also that iteration limits makes it
easier to profile and limit the time of execution.