Added the c++ database interface
The MalDB class in MalDB.h / MalDB.cpp is a c++ database interface
similar to the python implementation. The main difference is that
the ids are not the ids stored in the database (which are hardly
ever needed, except when inserting information), but rather 0-based
indices. This makes it easier to to implement the ratings matrix
code.