Changed MemReader to use pysparse and indices
Changed the MemReader class to use the pysparse sparse matrix
class to store the ratings matrix (avoiding the need to use
dictionarites) and use 0-based indices instead of the UIDs
that the website provides (the latter needs is only necessary
when adding things to the database to resolve name clashes,
but the former makes constructing matrices easier).

NOTE: this commit breaks all current code using the MemReader
class. This is okay, as all code using the MemReader was in
for massive changes anyway.