Removed the first feature vector term as an implicit bias term.
When training and predicting ratings, the first element in the
feature/taste vectors is now not treated specially treated as
a bias term that is multiplied by some constant factor, it is now
just like any other element in the vector.

This was done because the systematic bias is important to center
the baseline predictor about the true rating. If the bias is
used as a feature vector element, it will be regularized and kept
minimal (which is even worse with more factors), which leads to
higher RMSE's